Adjustment for better visualization of the calculateAmountReservation function call code

+++ src/main/java/br/com/phss/model/service/CalculateAmountReservation.java
@@ -6,8 +6,8 @@ import java.time.temporal.ChronoUnit;
 
 public class CalculateAmountReservation {
     public static BigDecimal calculateAmountReservation(LocalDate checkInDate, LocalDate checkOutDate) {
-        int numberOfDays = (int) checkInDate.until(checkOutDate, ChronoUnit.DAYS);
+        BigDecimal numberOfDays = BigDecimal.valueOf(checkInDate.until(checkOutDate, ChronoUnit.DAYS));
         BigDecimal dailyRate = new BigDecimal("100.00");
-        return new BigDecimal(String.valueOf((dailyRate.multiply(BigDecimal.valueOf(numberOfDays)))));
+        return new BigDecimal(String.valueOf(dailyRate.multiply(numberOfDays)));
     }
 }
